[ISSUE-123] 회원가입 유효성 완료
- 모든 유효성 정규식으로 변경
- 회원가입 후 생일 자리에 가입날짜가 들어가게 함
